requirements.txt
pandas
numpy
geopandas
rasterio
ipykernel
openpyxl
xgboost
scipy
scikit-image
matplotlib
hydra-core
tqdm
pulearn
wandb
